requirements.txt
# Core dependencies
langchain>=0.0.267
langchain-huggingface>=0.0.6
langchain-openai>=0.0.2
langchain-community>=0.0.3
transformers>=4.30.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pydantic>=1.10.8
openai>=0.27.8

# UI
streamlit>=1.24.0

# Data processing
pandas>=2.0.2
scikit-learn>=1.2.2
torch>=2.0.1
tqdm>=4.65.0
python-Levenshtein>=0.21.0

# Configuration and utils
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional - uncomment if needed
# faiss-gpu>=1.7.4  # For GPU support
# flask>=2.3.2      # For API server
# gunicorn>=20.1.0  # For production deployment